+++ ChangeLog.md
@@ -7,7 +7,8 @@ All notable changes to this project will be documented in this file.
 
 ## [1.3.1] - 2025-07-xx
 
-- ...
+- Added a quarantine handler that allows to move/copy infected files to a 
+  quarantine directory